requirements.txt
tkinter
Pillow>=9.0.0
numpy>=1.20.0
opencv-python
matplotlib>=3.5.0
pandas>=1.3.0
geopandas>=0.10.0
shapely>=1.8.0